refactor(ui): move update surface to a sidebar card, not a header banner
Per review: keep it small but noticeable in the sidebar rather than a
full-width bar across the top.

- New timeline/UpdateCard.tsx: same check_update/install_update/update-progress
  logic, restyled as an accent card matching OverviewPanel's "Board cleanup
  available" CTA.
- Mounted at the top of OverviewPanel (the sidebar overview); reverted the
  LayoutBanners header-banner mount and removed the old UpdateBanner.tsx.
- Test updated + renamed (update-card.test.ts): asserts the OverviewPanel
  mount, that it's NOT in LayoutBanners, command reuse, consent-based render,
  Tauri-safe degradation, and the card styling.
